+++ src/Position.h
@@ -3,7 +3,7 @@
 class Position
 {
 public:
-	Position(size_t rows, size_t columns);
-	size_t m_Rows;
-	size_t m_Columns;
+	Position(int rows, int columns);
+	int m_Rows;
+	int m_Columns;
 };
